requirements/base.txt
psycopg2-binary==2.9.10
python-environ==0.4.54
drf-yasg==1.21.8
cryptography==41.0.7
celery==5.3.6
redis==5.0.1
flower==2.0.1
firebase-admin==6.5.0
Django==5.1.3
django-cors-headers==4.6.0
django-filter==24.3
django-modeltranslation==0.19.10
django-redis==5.4.0
djangorestframework==3.15.2
import_export==0.3.1
django-import-export==3.3.1
django-jazzmin==3.0.1
django-recaptcha==3.0.0
djangorestframework-simplejwt==5.3.0
pillow==11.0.0
django-rosetta==0.9.9